Move parameter sources to the parameter_sources directory
