+++ solcore/parameter_sources/__init__.py
@@ -1,0 +1,2 @@
+from .simple_parameters import SimpleSource
+from .calculable_parameters import CalculableParameters
