+++ solcore/parameter_sources/simple_parameters.py
