+++ solcore/parameter_system/__init__.py
@@ -1,3 +1,1 @@
 from .parameter_system import ParameterSystem
-from .simple_parameters import SimpleSource
-from .calculable_parameters import CalculableParameters
